Lots of changes and lots of fixes that i cant describe in a commit message

+++ src/main/java/mc/warp/transmitgenerators/gson/GeneratorSerializer.kt
@@ -29,8 +29,8 @@ class GeneratorSerializer  : JsonSerializer<Generator>, JsonDeserializer<Generat
 
 
         var id = json.get("id").asString!!
-        var worth = json.get("sellValue").asInt!!
-        var price = json.get("upgradePrice").asInt!!
+        var worth = json.get("sellValue").asLong!!
+        var price = json.get("upgradePrice").asLong!!
         var upgrade = json.get("upgradeId").asString!!
